requirements.txt
PySocks==1.7.1
yandexfreetranslate
python-telegram-bot[ext]
